Add forwarding list to configuration. Remove unused classes.
