+++ config/src/main/java/com/quorum/tessera/config/adapters/PathAdapter.java
@@ -3,7 +3,6 @@ package com.quorum.tessera.config.adapters;
 import javax.xml.bind.annotation.adapters.XmlAdapter;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Objects;
 
 /**
  * Converts a String to a Path object for JAXB de/serialisation
@@ -12,20 +11,12 @@ public class PathAdapter extends XmlAdapter<String,Path> {
 
     @Override
     public Path unmarshal(final String input) {
-        if(Objects.isNull(input)) {
-            return null;
-        }
-
-        return Paths.get(input);
+        return (input == null) ? null : Paths.get(input);
     }
 
     @Override
     public String marshal(final Path input) {
-        if(Objects.isNull(input)) {
-            return null;
-        }
-
-        return input.toAbsolutePath().toString();
+        return (input == null) ? null : input.toAbsolutePath().toString();
     }
 
 }
