--- config/src/main/java/com/quorum/tessera/config/keys/KeyGeneratorException.java
@@ -1,9 +1,0 @@
-package com.quorum.tessera.config.keys;
-
-public class KeyGeneratorException extends RuntimeException {
-
-    public KeyGeneratorException(final Throwable cause) {
-        super(cause);
-    }
-
-}
